refactor: Remove transaction::util
Added test for taproot builder with scripts
Refactored taproot builder with scripts for better readability
